Added connector specific completion capabilities. Moved extensions and schemas (only postgres).
