+++ prisma-fmt/src/text_document_completion/datasource.rs
@@ -3,27 +3,9 @@ use std::collections::HashMap;
 use lsp_types::{
     CompletionItem, CompletionItemKind, CompletionList, Documentation, InsertTextFormat, MarkupContent, MarkupKind,
 };
+use psl::datamodel_connector::format_completion_docs;
 
-use super::{add_quotes, generate_pretty_doc, CompletionContext};
-
-pub(super) fn schemas_completion(completion_list: &mut CompletionList) {
-    completion_list.items.push(CompletionItem {
-        label: "schemas".to_owned(),
-        insert_text: Some(r#"schemas = [$0]"#.to_owned()),
-        insert_text_format: Some(InsertTextFormat::SNIPPET),
-        kind: Some(CompletionItemKind::FIELD),
-        documentation: Some(Documentation::MarkupContent(MarkupContent {
-            kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
-                r#"schemas = ["foo", "bar", "baz"]"#,
-                "The list of database schemas. [Learn More](https://pris.ly/d/multi-schema-configuration)",
-                None,
-            ),
-        })),
-        // detail: Some("schemas".to_owned()),
-        ..Default::default()
-    });
-}
+use super::{add_quotes, CompletionContext};
 
 pub(super) fn relation_mode_completion(completion_list: &mut CompletionList) {
     completion_list.items.push(CompletionItem {
@@ -33,7 +15,7 @@ pub(super) fn relation_mode_completion(completion_list: &mut CompletionList) {
         kind: Some(CompletionItemKind::FIELD),
         documentation: Some(Documentation::MarkupContent(MarkupContent {
             kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
+            value: format_completion_docs(
                 r#"relationMode = "foreignKeys" | "prisma""#,
                 r#"Set the global relation mode for all relations. Values can be either "foreignKeys" (Default), or "prisma". [Learn more](https://pris.ly/d/relation-mode)"#,
                 None,
@@ -51,7 +33,7 @@ pub(super) fn direct_url_completion(completion_list: &mut CompletionList) {
         kind: Some(CompletionItemKind::FIELD),
         documentation: Some(Documentation::MarkupContent(MarkupContent {
             kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
+            value: format_completion_docs(
                 r#"directUrl = "String" | env("ENVIRONMENT_VARIABLE")"#,
                 r#"Connection URL for direct connection to the database. [Learn more](https://pris.ly/d/data-proxy-cli)."#,
                 None,
@@ -69,7 +51,7 @@ pub(super) fn shadow_db_completion(completion_list: &mut CompletionList) {
         kind: Some(CompletionItemKind::FIELD),
         documentation: Some(Documentation::MarkupContent(MarkupContent {
             kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
+            value: format_completion_docs(
                 r#"shadowDatabaseUrl = "String" | env("ENVIRONMENT_VARIABLE")"#,
                 r#"Connection URL including authentication info to use for Migrate's [shadow database](https://pris.ly/d/migrate-shadow)."#,
                 None,
@@ -87,7 +69,7 @@ pub(super) fn url_completion(completion_list: &mut CompletionList) {
         kind: Some(CompletionItemKind::FIELD),
         documentation: Some(Documentation::MarkupContent(MarkupContent {
             kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
+            value: format_completion_docs(
                 r#"url = "String" | env("ENVIRONMENT_VARIABLE")"#,
                 r#"Connection URL including authentication info. Each datasource provider documents the URL syntax. Most providers use the syntax provided by the database. [Learn more](https://pris.ly/d/connection-strings)."#,
                 None,
@@ -105,27 +87,9 @@ pub(super) fn provider_completion(completion_list: &mut CompletionList) {
         kind: Some(CompletionItemKind::FIELD),
         documentation: Some(Documentation::MarkupContent(MarkupContent {
             kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
+            value: format_completion_docs(
                 r#"provider = "foo""#,
                 r#"Describes which datasource connector to use. Can be one of the following datasource providers: `postgresql`, `mysql`, `sqlserver`, `sqlite`, `mongodb` or `cockroachdb`."#,
-                None,
-            ),
-        })),
-        ..Default::default()
-    })
-}
-
-pub(super) fn extensions_completion(completion_list: &mut CompletionList) {
-    completion_list.items.push(CompletionItem {
-        label: "extensions".to_owned(),
-        insert_text: Some("extensions = [$0]".to_owned()),
-        insert_text_format: Some(InsertTextFormat::SNIPPET),
-        kind: Some(CompletionItemKind::FIELD),
-        documentation: Some(Documentation::MarkupContent(MarkupContent {
-            kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
-                r#"extensions = [pg_trgm, postgis(version: "2.1")]"#,
-                r#"Enable PostgreSQL extensions. [Learn more](https://pris.ly/d/postgresql-extensions)"#,
                 None,
             ),
         })),
@@ -141,7 +105,7 @@ pub(super) fn url_env_completion(completion_list: &mut CompletionList) {
         kind: Some(CompletionItemKind::PROPERTY),
         documentation: Some(Documentation::MarkupContent(MarkupContent {
             kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
+            value: format_completion_docs(
                 r#"env(_ environmentVariable: string)"#,
                 r#"Specifies a datasource via an environment variable. When running a Prisma CLI command that needs the database connection URL (e.g. `prisma db pull`), you need to make sure that the `DATABASE_URL` environment variable is set. One way to do so is by creating a `.env` file. Note that the file must be in the same directory as your schema.prisma file to automatically be picked up by the Prisma CLI.""#,
                 Some(HashMap::from([(
@@ -162,7 +126,7 @@ pub(super) fn url_quotes_completion(completion_list: &mut CompletionList) {
         kind: Some(CompletionItemKind::PROPERTY),
         documentation: Some(Documentation::MarkupContent(MarkupContent {
             kind: MarkupKind::Markdown,
-            value: generate_pretty_doc(
+            value: format_completion_docs(
                 r#""connectionString""#,
                 r#"Connection URL including authentication info. Each datasource provider documents the URL syntax. Most providers use the syntax provided by the database. [Learn more](https://pris.ly/d/prisma-schema)."#,
                 None,
