+++ psl/psl-core/src/datamodel_connector.rs
@@ -7,6 +7,9 @@ pub mod constraint_names;
 /// Extensions for parser database walkers with context from the connector.
 pub mod walker_ext_traits;
 
+/// Connector completions
+pub mod completions;
+
 mod empty_connector;
 mod filters;
 mod native_types;
@@ -14,13 +17,14 @@ mod relation_mode;
 
 pub use self::{
     capabilities::{ConnectorCapabilities, ConnectorCapability},
+    completions::format_completion_docs,
     empty_connector::EmptyDatamodelConnector,
     filters::*,
     native_types::{NativeTypeArguments, NativeTypeConstructor, NativeTypeInstance},
     relation_mode::RelationMode,
 };
 
-use crate::{configuration::DatasourceConnectorData, Datasource, PreviewFeature};
+use crate::{configuration::DatasourceConnectorData, Configuration, Datasource, PreviewFeature};
 use diagnostics::{DatamodelError, Diagnostics, NativeTypeErrorFactory, Span};
 use enumflags2::BitFlags;
 use lsp_types::CompletionList;
@@ -333,6 +337,8 @@ pub trait Connector: Send + Sync {
 
     fn push_completions(&self, _db: &ParserDatabase, _position: SchemaPosition<'_>, _completions: &mut CompletionList) {
     }
+
+    fn datasource_completions(&self, _config: &Configuration, _completions: &mut CompletionList) {}
 
     fn parse_datasource_properties(
         &self,
